# implement speedups with rust v2