--- a/.gitignore
+++ b/.gitignore
@@ -11,3 +11,4 @@ __pycache__/
 /.coverage*
 /htmlcov/
 /docs/_build/
+/src/rust/target/