--- a/src/markupsafe/_native.py
+++ b/src/markupsafe/_native.py
@@ -1,11 +1,11 @@
 from __future__ import annotations
 
-import typing as t
 
 def escape_inner(s: str, /) -> str:
-    return s
-        .replace("&", "&amp;")
+    return (
+        s.replace("&", "&amp;")
         .replace(">", "&gt;")
         .replace("<", "&lt;")
         .replace("'", "&#39;")
         .replace('"', "&#34;")
+    )