--- a/src/rust/Cargo.toml
+++ b/src/rust/Cargo.toml
@@ -4,6 +4,9 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
+[profile.release]
+debug = true
+
 [dependencies]
 pyo3 = "0.20.3"