… pyproject.toml
… src/markupsafe/_rust_speedups.pyi
… src/rust/Cargo.lock