@davidism 2024-04-23
I'll have to run it on my machine for an exact comparison, but those are some good performance numbers compared to the C. The `long escape` and `long plain` ratios are _significantly_ better, and the `long suffix` is equal rather than slower.