@carsonburr 2024-08-30
Simplified the rust speedups to use a lookup table instead of relying on auto-vectorized simd. The advantages of this are that it's not as messy, doesn't use advanced rust features, doesn't use unsafe*, and slightly faster for the workloads in bench.py. I'm sure this still could be simd-accelerated, but not portably.

*potentially eating the conversion cost if the PyString is stored as utf-16 or unicode 32-bit