@davidism 2024-10-06
@davidhewitt thanks for looking at this. In #461 we're setting up wheel builds for 313 and 313t (free threading). Say we were to add the `features = ["abi3"]` suggestion you made above. How would we deal with the free threading build? Presumably, we'd offer an abi3 wheel, and then also a 313t wheel that's not abi3? How would we configure pyo3 and cibuildwheel to handle both of those?