@davidhewitt 2024-10-07
For ease of use, currently at the moment if you have the `abi3` feature set but you're building on the freethreaded Python we'll ignore it and build for the freethreaded ABI. I understand in 3.14 there's possibility of a new stable abi which supports freethreading, so that might change in future.

I am not super familiar with cibuildwheel configuration 🙈, though I assume that the Rust build would work the same way as building a freethreaded C wheel.