@davidhewitt 2025-02-19
For what it's worth, we have now got good free-threading support in PyO3, and the abi3 / free-threaded interaction is as described; just build an abi3 wheel and a free-threaded wheel, and you're done.

See e.g. [support added in bcrypt](https://github.com/pyca/bcrypt/pull/955/files) which [also uses the abi3 feature](https://github.com/pyca/bcrypt/blob/77db0c3c85b77d5103a91f873ba0e46f8fa4a933/src/_bcrypt/Cargo.toml#L9).