@davidism 2025-02-19
Do you know if cibuildwheel is adding support as well? PyCA uses their own wheel building infrastructure, so it's not immediately clear how I would adapt their configuration.