@davidism 2025-02-19
https://cibuildwheel.pypa.io/en/stable/options Looks like there are various options for it.